extract AssetsExchangeRate;
